RubikCube is working properly (#18)

+++ src/Vector3D.java
@@ -92,11 +92,11 @@ public class Vector3D {
 
 	public void rotateVector(RotMatrix rotMatrix) {
 		int row = rotMatrix.mRotMatrix.length, col = rotMatrix.mRotMatrix[0].length;
-		double[] result = {0, 0, 0};
+		int[] result = {0, 0, 0};
 		
 		for (int i = 0; i < row; i++) {
 			for (int j = 0; j < col; j++) {
-				result[i] += (mVector[j] * rotMatrix.mRotMatrix[i][j]);
+				result[i] += (int) (mVector[j] * rotMatrix.mRotMatrix[i][j]);
 			}
 		}
 		mVector[0] = result[0];
